[text] draw input text to canvas
allows text entry and deletion in a subregion of the canvas. space,
tab, or enter to render a word; backspace to delete. supports
left-to-right scripts only for now.

the text rendering parameters work pretty well for the default serif
font in chrome on mac, but not necessarily in other browsers. next
TODO is to find a more portable approach.